+++ Sources/Views/ALKFriendGenericCardMessageCell.swift
@@ -32,6 +32,7 @@ open class ALKFriendGenericCardMessageCell: ALKGenericCardBaseCell {
         }
 
         static let maxWidth = UIScreen.main.bounds.width
+            - (ViewPadding.AvatarImageView.width + ViewPadding.AvatarImageView.leading)
         static let messageViewPadding = Padding(left: ChatCellPadding.ReceivedMessage.Message.left,
                                                 right: ChatCellPadding.ReceivedMessage.Message.right,
                                                 top: ChatCellPadding.ReceivedMessage.Message.top,
@@ -68,7 +69,7 @@ open class ALKFriendGenericCardMessageCell: ALKGenericCardBaseCell {
     fileprivate lazy var messageView = MessageView(
         bubbleStyle: MessageTheme.receivedMessage.bubble,
         messageStyle: MessageTheme.receivedMessage.message,
-        maxWidth: ViewPadding.maxWidth
+        maxWidth: ViewPadding.maxWidth - (ViewPadding.messageViewPadding.left + ViewPadding.messageViewPadding.right)
     )
     // fileprivate var messageViewPadding: Padding
     lazy var messageViewHeight = self.messageView.heightAnchor.constraint(equalToConstant: 0)